requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.0
opencv-python==4.5.1.48
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
SQLAlchemy==1.3.23
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
websockets==8.1
